Move files related to optimization to optimization/.
