+++ cartographer/mapping/internal/2d/pose_graph_2d_test.cc
@@ -137,7 +137,7 @@ class PoseGraph2DTest : public ::testing::Test {
       auto options = CreatePoseGraphOptions(parameter_dictionary.get());
       pose_graph_ = common::make_unique<PoseGraph2D>(
           options,
-          common::make_unique<pose_graph::OptimizationProblem2D>(
+          common::make_unique<optimization::OptimizationProblem2D>(
               options.optimization_problem_options()),
           &thread_pool_);
     }
