+++ cartographer/mapping/internal/3d/pose_graph_3d.cc
@@ -41,7 +41,7 @@ namespace mapping {
 
 PoseGraph3D::PoseGraph3D(
     const proto::PoseGraphOptions& options,
-    std::unique_ptr<pose_graph::OptimizationProblem3D> optimization_problem,
+    std::unique_ptr<optimization::OptimizationProblem3D> optimization_problem,
     common::ThreadPool* thread_pool)
     : options_(options),
       optimization_problem_(std::move(optimization_problem)),
@@ -267,7 +267,7 @@ void PoseGraph3D::ComputeConstraintsForNode(
       insertion_submaps.front()->local_pose().inverse() * local_pose;
   optimization_problem_->AddTrajectoryNode(
       matching_id.trajectory_id,
-      pose_graph::NodeSpec3D{constant_data->time, local_pose, global_pose});
+      optimization::NodeSpec3D{constant_data->time, local_pose, global_pose});
   for (size_t i = 0; i < insertion_submaps.size(); ++i) {
     const SubmapId submap_id = submap_ids[i];
     // Even if this was the last node added to 'submap_id', the submap will only
@@ -461,7 +461,7 @@ void PoseGraph3D::AddSubmapFromProto(
   submap_data_.at(submap_id).submap = submap_ptr;
   // Immediately show the submap at the 'global_submap_pose'.
   global_submap_poses_.Insert(submap_id,
-                              pose_graph::SubmapSpec3D{global_submap_pose});
+                              optimization::SubmapSpec3D{global_submap_pose});
   AddWorkItem([this, submap_id, global_submap_pose]() REQUIRES(mutex_) {
     submap_data_.at(submap_id).state = SubmapState::kFinished;
     optimization_problem_->InsertSubmap(submap_id, global_submap_pose);
@@ -483,8 +483,8 @@ void PoseGraph3D::AddNodeFromProto(const transform::Rigid3d& global_pose,
     const auto& constant_data = trajectory_nodes_.at(node_id).constant_data;
     optimization_problem_->InsertTrajectoryNode(
         node_id,
-        pose_graph::NodeSpec3D{constant_data->time, constant_data->local_pose,
-                               global_pose});
+        optimization::NodeSpec3D{constant_data->time, constant_data->local_pose,
+                                 global_pose});
   });
 }
 
@@ -778,7 +778,7 @@ PoseGraph3D::GetAllSubmapPoses() {
 }
 
 transform::Rigid3d PoseGraph3D::ComputeLocalToGlobalTransform(
-    const MapById<SubmapId, pose_graph::SubmapSpec3D>& global_submap_poses,
+    const MapById<SubmapId, optimization::SubmapSpec3D>& global_submap_poses,
     const int trajectory_id) const {
   auto begin_it = global_submap_poses.BeginOfTrajectory(trajectory_id);
   auto end_it = global_submap_poses.EndOfTrajectory(trajectory_id);
