+++ cartographer/mapping/pose_graph/proto/optimization_problem_options.proto
@@ -14,7 +14,7 @@
 
 syntax = "proto3";
 
-package cartographer.mapping.pose_graph.proto;
+package cartographer.mapping.optimization.proto;
 
 import "cartographer/common/proto/ceres_solver_options.proto";
